AccessColumn updated to use LRU Moved Vector into its own folder and made a partial class
Added AccessVal and AccessRow to VectorBase class
Removed AccessRow from Vector3
