+++ DataScience/Core/Vector/VectorToString.cs
@@ -1,0 +1,45 @@
+﻿using DataScience.Utility;
+using System;
+using System.Linq;
+using System.Text;
+
+namespace DataScience
+{
+    public partial class Vector
+    {
+        public override string ToString()
+        {
+            bool neg = (this.Value.Min() < 0);
+
+            int displace = new int[] { ((int)Max()).ToString().Length, ((int)Min()).ToString().Length }.Max();
+            int maxchar = $"{displace:0.00}".Length;
+
+            if (displace > maxchar)
+            {
+                int temp = displace;
+                displace = maxchar;
+                maxchar = temp;
+            }
+
+            StringBuilder stringBuilder = new StringBuilder();
+
+            for (int i = 0; i < this.Value.Length; i++)
+            {
+                if ((i % this.Columns == 0) && i != 0)
+                {
+                    stringBuilder.AppendLine();
+                }
+
+                string val = $"{this.Value[i]:0.00}";
+                int disp = displace - ((int)Math.Floor(MathF.Abs(this.Value[i]))).ToString().Length;
+
+                stringBuilder.AppendFormat($"| {Util.PadBoth(val, maxchar, disp, this.Value[i] < 0f)} |");
+            }
+
+            return stringBuilder.AppendLine().ToString();
+        }
+
+
+
+    }
+}
